requirements.txt
torch
torchvision
transformers
datasets
accelerate
numpy
pandas
matplotlib
scipy
tqdm
Pillow
